some work from home. Fuck you strncmp. Made you my bitch

+++ libft.h
@@ -35,7 +35,7 @@ char	*strrchr(const char *s, int c);
 /*strnstr*/
 char *ft_strncat(char *dest, const char *src, size_t n);
 int ft_strcmp(const char *s1, const char *s2);
-int ft_strncmp(const char *s1, const char *s2, int n);
+int ft_strncmp(const char *s1, const char *s2, size_t n);
 int ft_atoi(const char *str);
 /*memset*/
 /*bzero*/
